Isolate redis and memcached code
